adding artifacts to nlp model
